Change for download resource

+++ comunication/src/main/java/com/chuys/gshp/comunication/domain/model/MediaModel.kt
@@ -1,7 +1,7 @@
 package com.chuys.gshp.comunication.domain.model
 
 data class MediaModel(
-    var id: Long?,
+    var id: Int?,
     val title: String,
     val url: String,
     val description: String,
